Revert "Cambios"
This reverts commit 715a076b51ec2fa6c33173e0effc39b164650e1e, reversing
changes made to 3fed7e6fdc224974edab754642d7629925902609.

+++ src/main/java/com/Clubbr/Clubbr/Service/itemService.java
@@ -3,12 +3,17 @@ package com.Clubbr.Clubbr.Service;
 import com.Clubbr.Clubbr.Entity.item;
 import com.Clubbr.Clubbr.Entity.stablishment;
 import com.Clubbr.Clubbr.Entity.user;
-import com.Clubbr.Clubbr.advice.ResourceNotFoundException;
+import com.Clubbr.Clubbr.advice.ItemNotFoundException;
+import com.Clubbr.Clubbr.advice.ManagerNotFoundException;
+import com.Clubbr.Clubbr.advice.ManagerNotFromStablishmentException;
+import com.Clubbr.Clubbr.advice.StablishmentNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.Clubbr.Clubbr.Repository.stablishmentRepo;
 import com.Clubbr.Clubbr.Repository.itemRepo;
+import com.Clubbr.Clubbr.Repository.userRepo;
+import com.Clubbr.Clubbr.Repository.managerRepo;
 import com.Clubbr.Clubbr.Entity.manager;
 
 import java.util.List;
@@ -17,83 +22,116 @@ import java.util.List;
 public class itemService {
 
     @Autowired
-    private stablishmentService stabService;
+    private stablishmentRepo stabRepo;
 
     @Autowired
-    private stablishmentRepo stabRepo;
+    private stablishmentService stabService;
 
     @Autowired
     private itemRepo itemRepo;
 
     @Autowired
-    private userService userService;
+    private userRepo userRepo;
 
     @Autowired
     private jwtService jwtService;
 
     @Autowired
-    private managerService managerService;
+    private managerRepo managerRepo;
 
     @Transactional
     public void addItemToStablishment(Long stablishmentID, item newItem, String token){
-        stablishment targetStab = stabService.getStab(stablishmentID);
-        user user = userService.getUser(jwtService.extractUserIDFromToken(token));
+        stablishment stablishment = stabRepo.findById(stablishmentID).orElse(null);
+        user user = userRepo.findById(jwtService.extractUserIDFromToken(token)).orElse(null);
+        manager manager = managerRepo.findByUserID(user).orElse(null);
 
-        if (userService.isManager(user)){
-            manager targetManager = managerService.getManager(userService.getUser(jwtService.extractUserIDFromToken(token)));
-            if (!managerService.isManagerInStab(targetStab, targetManager)){
-                throw new ResourceNotFoundException("Manager", "userID", user.getUserID(), "Establecimiento", "stablishmentID", targetStab.getStablishmentID());
-            }
+        if (stablishment == null){
+            throw new StablishmentNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
+        }
+
+        if (manager == null){
+            throw new ManagerNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
+        }
+
+        if (!stabService.isManagerInStab(stablishment, manager)){
+            throw new ManagerNotFromStablishmentException("El usuario con el ID " + user.getUserID() + " no es manager del establecimiento con el ID " + stablishmentID);
         }
 
         newItem.setItemQuantity(newItem.getItemStock());
-        newItem.setStablishmentID(targetStab);
-        targetStab.getInventory().add(newItem);
-        stabRepo.save(targetStab);
+        newItem.setStablishmentID(stablishment);
+        stablishment.getInventory().add(newItem);
+        stabRepo.save(stablishment);
         itemRepo.save(newItem);
     }
 
     @Transactional(readOnly = true)
     public List<item> getItemsFromStablishment(Long stablishmentID, String token){
-        stablishment targetStab = stabService.getStab(stablishmentID);
-        user user = userService.getUser(jwtService.extractUserIDFromToken(token));
+        stablishment stablishment = stabRepo.findById(stablishmentID).orElse(null);
+        user user = userRepo.findById(jwtService.extractUserIDFromToken(token)).orElse(null);
+        manager manager = managerRepo.findByUserID(user).orElse(null);
 
-        if (userService.isManager(user)){
-            manager targetManager = managerService.getManager(userService.getUser(jwtService.extractUserIDFromToken(token)));
-            if (!managerService.isManagerInStab(targetStab, targetManager)){
-                throw new ResourceNotFoundException("Manager", "userID", user.getUserID(), "Establecimiento", "stablishmentID", targetStab.getStablishmentID());
-            }
+        if (stablishment == null){
+            throw new StablishmentNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
         }
-        return itemRepo.findByStablishmentID(targetStab);
+
+        if (manager == null){
+            throw new ManagerNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
+        }
+
+        if (!stabService.isManagerInStab(stablishment, manager)){
+            throw new ManagerNotFromStablishmentException("El usuario con el ID " + user.getUserID() + " no es manager del establecimiento con el ID " + stablishmentID);
+        }
+        return itemRepo.findByStablishmentID(stablishment);
     }
 
     @Transactional(readOnly = true)
     public item getItemFromStablishment(Long stablishmentID, Long itemID, String token){
-        stablishment targetStab = stabService.getStab(stablishmentID);
-        user user = userService.getUser(jwtService.extractUserIDFromToken(token));
+        item item = itemRepo.findById(itemID).orElse(null);
+        stablishment stablishment = stabRepo.findById(stablishmentID).orElse(null);
+        user user = userRepo.findById(jwtService.extractUserIDFromToken(token)).orElse(null);
+        manager manager = managerRepo.findByUserID(user).orElse(null);
 
-        if (userService.isManager(user)){
-            manager targetManager = managerService.getManager(userService.getUser(jwtService.extractUserIDFromToken(token)));
-            if (!managerService.isManagerInStab(targetStab, targetManager)){
-                throw new ResourceNotFoundException("Manager", "userID", user.getUserID(), "Establecimiento", "stablishmentID", targetStab.getStablishmentID());
-            }
+        if (stablishment == null){
+            throw new StablishmentNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
         }
 
-        item item = getItem(itemID);
-        if (item.getStablishmentID() != targetStab){
-            throw new ResourceNotFoundException("Item", "itemID", itemID, "Establecimiento", "stablishmentID", targetStab.getStablishmentID());
+        if (manager == null){
+            throw new ManagerNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
         }
-        
+
+        if (!stabService.isManagerInStab(stablishment, manager)){
+            throw new ManagerNotFromStablishmentException("El usuario con el ID " + user.getUserID() + " no es manager del establecimiento con el ID " + stablishmentID);
+        }
+
+        if (item == null || item.getStablishmentID() != stablishment){
+            throw new ItemNotFoundException("No se ha encontrado el item con el ID " + itemID + " en el establecimiento con el ID" + stablishmentID);
+        }
         return item;
-    }
-
-    public item getItem(Long itemID){
-        return itemRepo.findById(itemID).orElseThrow(() -> new ResourceNotFoundException("Item", "itemID", itemID));
     }
 
     @Transactional
     public void updateItemFromStablishment(Long stablishmentID, Long itemID, item updateItem, String token){
-        item item = getItemFromStablishment(stablishmentID, itemID, token);
+        stablishment stablishment = stabRepo.findById(stablishmentID).orElse(null);
+        item item = itemRepo.findById(itemID).orElse(null);
+
+        user user = userRepo.findById(jwtService.extractUserIDFromToken(token)).orElse(null);
+        manager manager = managerRepo.findByUserID(user).orElse(null);
+
+        if (stablishment == null){
+            throw new StablishmentNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
+        }
+
+        if (manager == null){
+            throw new ManagerNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
+        }
+
+        if (!stabService.isManagerInStab(stablishment, manager)){
+            throw new ManagerNotFromStablishmentException("El usuario con el ID " + user.getUserID() + " no es manager del establecimiento con el ID " + stablishmentID);
+        }
+
+        if (item == null || item.getStablishmentID() != stablishment){
+            throw new ItemNotFoundException("No se ha encontrado el item con el ID " + itemID + " en el establecimiento con el ID" + stablishmentID);
+        }
 
         item.setItemQuantity(updateItem.getItemQuantity());
         item.setItemDistributor(updateItem.getItemDistributor());
@@ -107,7 +145,27 @@ public class itemService {
 
     @Transactional
     public void deleteItemFromStablishment(Long stablishmentID, Long itemID, String token){
-        item item = getItemFromStablishment(stablishmentID, itemID, token);
+        stablishment stablishment = stabRepo.findById(stablishmentID).orElse(null);
+        item item = itemRepo.findById(itemID).orElse(null);
+
+        user user = userRepo.findById(jwtService.extractUserIDFromToken(token)).orElse(null);
+        manager manager = managerRepo.findByUserID(user).orElse(null);
+
+        if (stablishment == null){
+            throw new StablishmentNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
+        }
+
+        if (manager == null){
+            throw new ManagerNotFoundException("No se ha encontrado el establecimiento con el ID " + stablishmentID);
+        }
+
+        if (!stabService.isManagerInStab(stablishment, manager)){
+            throw new ManagerNotFromStablishmentException("El usuario con el ID " + user.getUserID() + " no es manager del establecimiento con el ID " + stablishmentID);
+        }
+
+        if (item == null || item.getStablishmentID() != stablishment){
+            throw new ItemNotFoundException("No se ha encontrado el item con el ID " + itemID + " en el establecimiento con el ID" + stablishmentID);
+        }
         itemRepo.delete(item);
     }
 }
